🏗️ Architect: Modularize apts.searches into a package structure
Refactor monolithic apts/searches.py into a modular apts/searches/ package:
- Extract _get_observer helper into apts/searches/utils.py
- Extract extrema search wrappers into apts/searches/extrema.py
- Extract conjunction and occultation search wrappers into apts/searches/conjunctions.py
- Re-export all public API functions in apts/searches/__init__.py for 100% backward compatibility
