+++ apts/searches/__init__.py
@@ -1,0 +1,21 @@
+from .conjunctions import (
+    find_conjunctions,
+    find_lunar_occultations,
+    find_mercury_inferior_conjunctions,
+)
+from .extrema import (
+    find_aphelion_perihelion,
+    find_highest_altitude,
+    find_moon_apogee_perigee,
+)
+from .utils import _get_observer
+
+__all__ = [
+    "_get_observer",
+    "find_highest_altitude",
+    "find_aphelion_perihelion",
+    "find_moon_apogee_perigee",
+    "find_conjunctions",
+    "find_mercury_inferior_conjunctions",
+    "find_lunar_occultations",
+]
